+++ package-arbor/bin/run_arbor_pipeline.pl
@@ -3,7 +3,7 @@
 use strict;
 use warnings;
 use Getopt::Long qw(:config no_ignore_case no_auto_abbrev pass_through);
-use lib ("/usr/local/projects/ergatis/package-rnaseq/lib/perl5");
+use lib ("/usr/local/projects/ergatis/package-arbor/lib/perl5");
 use Ergatis::Pipeline;
 use Ergatis::SavedPipeline;
 use Ergatis::ConfigFile;
@@ -12,7 +12,7 @@ my %options;
 my $results = GetOptions (\%options,
                           "layout|l=s",
                           "config|c=s",
-                          "ergatis_config|e=s",
+                          "ergatis_ini|e=s",
                           "repository_root|r=s",
                           "email_id|m=s"
                           );
@@ -27,14 +27,14 @@ my $config = $options{'config'};
 my $user_email = "";
 $user_email = $options{'email_id'} if( $options{'email_id'} );
 my $ecfg; 
-$ecfg = $options{'ergatis_config'} if( $options{'ergatis_config'} );
+$ecfg = $options{'ergatis_ini'} if( $options{'ergatis_ini'} );
 
-my $ergatis_config;
+my $ergatis_ini;
 if( $ecfg ) {
-    $ergatis_config = new Ergatis::ConfigFile( '-file' => $ecfg );
+    $ergatis_ini = new Ergatis::ConfigFile( '-file' => $ecfg );
 }
 
-my $id = &make_pipeline( $layout, $repo_root, $id_repo, $config, $ergatis_config );
+my $id = &make_pipeline( $layout, $repo_root, $id_repo, $config, $ergatis_ini );
 
 # Label the pipeline.
 # &label_pipeline( $repo_root, $id);
@@ -43,36 +43,23 @@ my $url = "http://ergatis.igs.umaryland.edu/cgi/view_pipeline.cgi?instance=$repo
 print "pipeline_id -> $id | pipeline_url -> $url\n";
 
 sub make_pipeline {
-    my ($pipeline_layout, $repository_root, $id_repo, $config, $ergatis_config) = @_;
+    my ($pipeline_layout, $repository_root, $id_repo, $config, $ergatis_ini) = @_;
     my $template = new Ergatis::SavedPipeline( 'template' => $pipeline_layout );
     $template->configure_saved_pipeline( $config, $repository_root, $id_repo );
     my $pipeline_id = $template->pipeline_id();    
-    if( $ergatis_config ) {
+    if( $ergatis_ini ) {
         my $xml = $repository_root."/workflow/runtime/pipeline/$pipeline_id/pipeline.xml";
         my $pipeline = new Ergatis::Pipeline( id => $pipeline_id,
                                               path => $xml );
         if(length($user_email) > 0) {
-        	$pipeline->run( 'ergatis_cfg' => $ergatis_config, 'email_user' => $user_email );
+        	$pipeline->run( 'ergatis_cfg' => $ergatis_ini, 'email_user' => $user_email );
         }
         else {
-        	$pipeline->run( 'ergatis_cfg' => $ergatis_config );
+        	$pipeline->run( 'ergatis_cfg' => $ergatis_ini );
         }
     }
     return $pipeline_id;
 }
-
-# sub label_pipeline {
-# 	my ($repository_root, $pipeline_id) = @_; 
-# 	my $pipeline_runtime_folder = "$repository_root/workflow/runtime/pipeline/$pipeline_id";
-# 	my $groups_file = "$pipeline_runtime_folder/pipeline.xml.groups";
-#         open( GROUP, ">> $groups_file") or die("Unable to open $groups_file for writing ($!)");
-#         print GROUP "$comment\n";
-#         close(GROUP);
-# 	my $comment_file = "$pipeline_runtime_folder/pipeline.xml.comment";
-#         open( COM, ">> $comment_file") or die("Unable to open $comment_file for writing ($!)");
-#         print COM "$comment";
-#         close(COM);
-# }
 
 sub check_options {
     my ($opts) = @_;