+++ package-arbor/CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+## 4/6/16
+* Wrote first draft of the pipeline creation scripts
+
 ## 4/4/16
 * Renamed the NICU branch to 'arbor'
 * Committed a bunch of scripts for the pipeline